osd/PrimaryLogPG: move BackfillInterval to recovery_types.
It will be needed in crimson as well.
